Fixed driver DTO, service mismatch, ID datatype issue, and driver deliveries API

+++ src/main/java/com/foodservice/controller/DeliveryDriverController.java
@@ -20,7 +20,7 @@ public class DeliveryDriverController {
 
     // ------- Get Driver by ID ----------
     @GetMapping("/{driverId}")
-    public ResponseEntity<DeliveryDriverDTO> getDriverById(@PathVariable Long driverId) {
+    public ResponseEntity<DeliveryDriverDTO> getDriverById(@PathVariable Integer driverId) {
         return ResponseEntity.ok(deliveryDriverservice.getDriverById(driverId));
     }
